move logic outisde of the viewmodel
