+++ app/src/test/kotlin/com/wire/android/ui/home/conversations/ConversationsViewModelTest.kt
@@ -1,65 +1,16 @@
 package com.wire.android.ui.home.conversations
 
-import android.content.Context
-import androidx.lifecycle.SavedStateHandle
 import com.wire.android.config.CoroutineTestExtension
-import com.wire.android.config.TestDispatcherProvider
-import com.wire.android.mapper.MessageMapper
-import com.wire.android.mapper.UserTypeMapper
-import com.wire.android.model.UserStatus
-import com.wire.android.navigation.NavigationManager
 import com.wire.android.ui.home.conversations.ConversationViewModel.Companion.ASSET_SIZE_DEFAULT_LIMIT_BYTES
 import com.wire.android.ui.home.conversations.ConversationViewModel.Companion.IMAGE_SIZE_LIMIT_BYTES
 import com.wire.android.ui.home.conversations.model.AttachmentBundle
 import com.wire.android.ui.home.conversations.model.AttachmentType
-import com.wire.android.ui.home.conversations.model.MessageHeader
-import com.wire.android.ui.home.conversations.model.MessageSource
-import com.wire.android.ui.home.conversations.model.MessageStatus
-import com.wire.android.ui.home.conversations.model.UIMessage
-import com.wire.android.ui.home.conversations.model.User
-import com.wire.android.ui.home.conversationslist.model.Membership
-import com.wire.android.util.FileManager
-import com.wire.android.util.ui.UIText
-import com.wire.kalium.logic.data.conversation.ClientId
-import com.wire.kalium.logic.data.conversation.Conversation
-import com.wire.kalium.logic.data.conversation.ConversationDetails
-import com.wire.kalium.logic.data.conversation.LegalHoldStatus
-import com.wire.kalium.logic.data.conversation.MemberDetails
-import com.wire.kalium.logic.data.conversation.UserType
-import com.wire.kalium.logic.data.id.ConversationId
-import com.wire.kalium.logic.data.message.Message
-import com.wire.kalium.logic.data.message.MessageContent.Text
-import com.wire.kalium.logic.data.publicuser.model.OtherUser
 import com.wire.kalium.logic.data.team.Team
-import com.wire.kalium.logic.data.user.ConnectionState
-import com.wire.kalium.logic.data.user.SelfUser
-import com.wire.kalium.logic.data.user.UserAssetId
 import com.wire.kalium.logic.data.user.UserId
-import com.wire.kalium.logic.feature.asset.GetMessageAssetUseCase
-import com.wire.kalium.logic.feature.asset.SendAssetMessageResult
-import com.wire.kalium.logic.feature.asset.SendAssetMessageUseCase
-import com.wire.kalium.logic.feature.asset.SendImageMessageResult
-import com.wire.kalium.logic.feature.asset.SendImageMessageUseCase
-import com.wire.kalium.logic.feature.asset.UpdateAssetMessageDownloadStatusUseCase
-import com.wire.kalium.logic.feature.conversation.ObserveConversationDetailsUseCase
-import com.wire.kalium.logic.feature.conversation.ObserveConversationMembersUseCase
-import com.wire.kalium.logic.feature.message.DeleteMessageUseCase
-import com.wire.kalium.logic.feature.message.GetRecentMessagesUseCase
-import com.wire.kalium.logic.feature.message.MarkMessagesAsNotifiedUseCase
-import com.wire.kalium.logic.feature.message.Result.Success
-import com.wire.kalium.logic.feature.message.SendTextMessageUseCase
-import com.wire.kalium.logic.feature.team.GetSelfTeamUseCase
-import io.mockk.MockKAnnotations
-import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
-import io.mockk.impl.annotations.MockK
-import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.consumeAsFlow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
 import org.amshove.kluent.internal.assertEquals
 import org.amshove.kluent.internal.assertFalse
@@ -133,23 +84,21 @@ class ConversationsViewModelTest {
     @Test
     fun `given a 1 on 1 conversation, when solving the conversation name, then the name of the other user is used`() = runTest {
         // Given
-        val conversationDetails = withMockConversationDetailsOneOnOne("Other User Name Goes Here")
-        val otherUserName = conversationDetails.otherUser.name
-        val (_, viewModel) = Arrangement().withChannelUpdates(conversationDetails = conversationDetails).arrange()
+        val oneToOneConversationDetails = withMockConversationDetailsOneOnOne("Other User Name Goes Here")
+        val (_, viewModel) = Arrangement().withConversationDetailUpdate(conversationDetails = oneToOneConversationDetails).arrange()
 
         // When - Then
-        assertEquals(otherUserName, viewModel.conversationViewState.conversationName)
+        assertEquals(oneToOneConversationDetails.otherUser.name, viewModel.conversationViewState.conversationName)
     }
 
     @Test
     fun `given a group conversation, when solving the conversation name, then the name of the conversation is used`() = runTest {
         // Given
-        val conversationDetails = mockConversationDetailsGroup("Conversation Name Goes Here")
-        val conversationName = conversationDetails.conversation.name
-        val (_, viewModel) = Arrangement().withChannelUpdates(conversationDetails = conversationDetails).arrange()
+        val groupConversationDetails = mockConversationDetailsGroup("Conversation Name Goes Here")
+        val (_, viewModel) = Arrangement().withConversationDetailUpdate(conversationDetails = groupConversationDetails).arrange()
 
         // When - Then
-        assertEquals(conversationName, viewModel.conversationViewState.conversationName)
+        assertEquals( groupConversationDetails.conversation.name, viewModel.conversationViewState.conversationName)
     }
 
     @Test
@@ -157,24 +106,23 @@ class ConversationsViewModelTest {
         // Given
         val firstConversationDetails = mockConversationDetailsGroup("Conversation Name Goes Here")
         val secondConversationDetails = mockConversationDetailsGroup("Conversation Name Was Updated")
-        val (arrangement, viewModel) = Arrangement().withChannelUpdates(conversationDetails = firstConversationDetails).arrange()
+        val (arrangement, viewModel) = Arrangement().withConversationDetailUpdate(conversationDetails = firstConversationDetails).arrange()
 
         // When - Then
         assertEquals(firstConversationDetails.conversation.name, viewModel.conversationViewState.conversationName)
 
         // When - Then
-        arrangement.withChannelUpdates(conversationDetails = secondConversationDetails)
+        arrangement.withConversationDetailUpdate(conversationDetails = secondConversationDetails)
         assertEquals(secondConversationDetails.conversation.name, viewModel.conversationViewState.conversationName)
     }
-
     @Test
     fun `given message sent by self user, when solving the message header, then the state should contain the self user name`() = runTest {
         // Given
-        val senderId = UserId("value", "domain")
-        val messages = listOf(mockedMessage(senderId = senderId))
         val selfUserName = "self user"
-        val selfMember = mockSelfUserDetails(selfUserName, senderId)
-        val (arrangement, viewModel) = Arrangement().withChannelUpdates(messages, listOf(selfMember)).arrange()
+        val messages = listOf(mockUIMessage(selfUserName))
+        val (arrangement, viewModel) = Arrangement()
+            .withMessagesUpdate(messages)
+            .arrange()
 
         // When - Then
         every { arrangement.uiText.asString(any()) } returns (selfUserName)
@@ -184,16 +132,11 @@ class ConversationsViewModelTest {
     @Test
     fun `given message sent by another user, when solving the message header, then the state should contain that user name`() = runTest {
         // Given
-        val senderId = UserId("value", "domain")
-        val messages = listOf(mockedMessage(senderId = senderId))
         val otherUserName = "other user"
-
-
-        val otherMember = mockOtherUserDetails(otherUserName, senderId)
-        val (arrangement, viewModel) = Arrangement().withChannelUpdates(messages, listOf(otherMember)).arrange()
-        coEvery { arrangement.messageMapper.toUIMessages(any(), any()) } returns listOf(
-            mockUIMessage(otherUserName)
-        )
+        val messages = listOf(mockUIMessage(otherUserName))
+        val (arrangement, viewModel) = Arrangement()
+            .withMessagesUpdate(messages)
+            .arrange()
 
         // When - Then
         every { arrangement.uiText.asString(any()) } returns (otherUserName)
@@ -203,11 +146,12 @@ class ConversationsViewModelTest {
     @Test
     fun `given the sender is updated, when solving the message header, then the update is propagated in the state`() = runTest {
         // Given
-        val senderId = UserId("value", "domain")
-        val messages = listOf(mockedMessage(senderId))
         val firstUserName = "other user"
+        val originalMessages = listOf(mockUIMessage(firstUserName))
         val secondUserName = "User changed their name"
-        val (arrangement, viewModel) = Arrangement().withChannelUpdates(messages, listOf(mockOtherUserDetails(firstUserName, senderId)))
+        val updatedMessages = listOf(mockUIMessage(secondUserName))
+        val (arrangement, viewModel) = Arrangement()
+            .withMessagesUpdate(originalMessages)
             .arrange()
 
         // When - Then
@@ -216,7 +160,10 @@ class ConversationsViewModelTest {
 
         // When - Then
         every { arrangement.uiText.asString(any()) } returns (secondUserName)
-        arrangement.withChannelUpdates(messages, listOf(mockOtherUserDetails(secondUserName, senderId)))
+        arrangement
+            .withMessagesUpdate(updatedMessages)
+            .arrange()
+
         assertEquals(secondUserName, viewModel.conversationViewState.messages.first().messageHeader.username.asString(arrangement.context))
     }
 
@@ -268,7 +215,7 @@ class ConversationsViewModelTest {
         // Given
         val conversationDetails = withMockConversationDetailsOneOnOne("", "userAssetId")
         val otherUserAvatar = conversationDetails.otherUser.previewPicture
-        val (_, viewModel) = Arrangement().withChannelUpdates(conversationDetails = conversationDetails).arrange()
+        val (_, viewModel) = Arrangement().withConversationDetailUpdate(conversationDetails = conversationDetails).arrange()
         val actualAvatar = viewModel.conversationViewState.conversationAvatar
         // When - Then
         assert(actualAvatar is ConversationAvatar.OneOne)
@@ -368,210 +315,4 @@ class ConversationsViewModelTest {
             assert(viewModel.conversationViewState.downloadedAssetDialogState == DownloadedAssetDialogVisibilityState.Hidden)
         }
 
-    private class Arrangement {
-        init {
-            // Tests setup
-            val dummyConversationId = "some-dummy-value@some.dummy.domain"
-            MockKAnnotations.init(this, relaxUnitFun = true)
-            every { savedStateHandle.get<String>(any()) } returns dummyConversationId
-            every { savedStateHandle.set(any(), any<String>()) } returns Unit
-
-            // Default empty values
-            coEvery { getMessages(any()) } returns flowOf(listOf())
-            coEvery { observeMemberDetails(any()) } returns flowOf(listOf())
-            coEvery { observeConversationDetails(any()) } returns flowOf()
-            coEvery { markMessagesAsNotified(any(), any()) } returns Success
-            coEvery { getSelfUserTeam() } returns flowOf()
-            coEvery { messageMapper.toUIMessages(any(), any()) } returns listOf()
-        }
-
-        @MockK
-        private lateinit var savedStateHandle: SavedStateHandle
-
-        @MockK
-        lateinit var navigationManager: NavigationManager
-
-        @MockK
-        lateinit var getMessages: GetRecentMessagesUseCase
-
-        @MockK
-        lateinit var sendTextMessage: SendTextMessageUseCase
-
-        @MockK
-        lateinit var sendAssetMessage: SendAssetMessageUseCase
-
-        @MockK
-        lateinit var sendImageMessage: SendImageMessageUseCase
-
-        @MockK
-        lateinit var getMessageAsset: GetMessageAssetUseCase
-
-        @MockK
-        lateinit var deleteMessage: DeleteMessageUseCase
-
-        @MockK
-        lateinit var observeConversationDetails: ObserveConversationDetailsUseCase
-
-        @MockK
-        lateinit var observeMemberDetails: ObserveConversationMembersUseCase
-
-        @MockK
-        lateinit var markMessagesAsNotified: MarkMessagesAsNotifiedUseCase
-
-        @MockK
-        lateinit var updateAssetMessageDownloadStatus: UpdateAssetMessageDownloadStatusUseCase
-
-        @MockK
-        lateinit var getSelfUserTeam: GetSelfTeamUseCase
-
-        @MockK
-        lateinit var fileManager: FileManager
-
-        @MockK
-        lateinit var messageMapper: MessageMapper
-
-        @MockK
-        lateinit var context: Context
-
-        @MockK
-        lateinit var uiText: UIText
-
-        val otherMemberUpdatesChannel = Channel<List<MemberDetails>>(capacity = Channel.UNLIMITED)
-        val conversationDetailsChannel = Channel<ConversationDetails>(capacity = Channel.UNLIMITED)
-
-
-        private val viewModel by lazy {
-            ConversationViewModel(
-                savedStateHandle = savedStateHandle,
-                navigationManager = navigationManager,
-                getMessages = getMessages,
-                observeConversationDetails = observeConversationDetails,
-                observeMemberDetails = observeMemberDetails,
-                sendTextMessage = sendTextMessage,
-                sendAssetMessage = sendAssetMessage,
-                sendImageMessage = sendImageMessage,
-                getMessageAsset = getMessageAsset,
-                deleteMessage = deleteMessage,
-                dispatchers = TestDispatcherProvider(),
-                markMessagesAsNotified = markMessagesAsNotified,
-                updateAssetMessageDownloadStatus = updateAssetMessageDownloadStatus,
-                getSelfUserTeam = getSelfUserTeam,
-                fileManager = fileManager,
-                messageMapper = messageMapper
-            )
-        }
-
-        suspend fun withChannelUpdates(
-            messages: List<Message> = emptyList(),
-            members: List<MemberDetails> = emptyList(),
-            conversationDetails: ConversationDetails? = null
-        ): Arrangement {
-            coEvery { getMessages(any()) } returns flowOf(messages)
-            coEvery { observeMemberDetails(any()) } returns otherMemberUpdatesChannel.consumeAsFlow()
-            coEvery { observeConversationDetails(any()) } returns conversationDetailsChannel.consumeAsFlow()
-
-            otherMemberUpdatesChannel.send(members)
-            conversationDetails?.run { conversationDetailsChannel.send(this) }
-            return this
-        }
-
-        fun withSuccessfulSendAttachmentMessage(): Arrangement {
-            coEvery { sendAssetMessage(any(), any(), any(), any()) } returns SendAssetMessageResult.Success
-            coEvery { sendImageMessage(any(), any(), any(), any(), any()) } returns SendImageMessageResult.Success
-            return this
-        }
-
-        fun withSuccessfulSaveAssetMessage(assetName: String, assetData: ByteArray, messageId: String): Arrangement {
-            viewModel.showOnAssetDownloadedDialog(assetName, assetData, messageId)
-            coEvery { fileManager.saveToExternalStorage(any(), any(), any()) }.answers {
-                viewModel.hideOnAssetDownloadedDialog()
-            }
-            return this
-        }
-
-        fun withSuccessfulOpenAssetMessage(assetName: String, assetData: ByteArray, messageId: String): Arrangement {
-            viewModel.showOnAssetDownloadedDialog(assetName, assetData, messageId)
-            every { fileManager.openWithExternalApp(any(), any(), any()) }.answers {
-                viewModel.hideOnAssetDownloadedDialog()
-            }
-            return this
-        }
-
-        fun withTeamUser(userTeam: Team): Arrangement {
-            coEvery { getSelfUserTeam() } returns flowOf(userTeam)
-            return this
-        }
-
-        fun arrange() = this to viewModel
-
-    }
-
-    private fun withMockConversationDetailsOneOnOne(senderName: String, senderAvatar: UserAssetId? = null) = ConversationDetails.OneOne(
-        mockk(),
-        mockk<OtherUser>().apply {
-            every { name } returns senderName
-            every { previewPicture } returns senderAvatar
-        },
-        ConnectionState.PENDING,
-        LegalHoldStatus.DISABLED,
-        UserType.INTERNAL
-    )
-
-    private fun mockConversationDetailsGroup(conversationName: String) = ConversationDetails.Group(mockk<Conversation>().apply {
-        every { name } returns conversationName
-        every { id } returns ConversationId("someId", "someDomain")
-    }, mockk())
-
-    private fun mockSelfUserDetails(
-        name: String,
-        id: UserId = UserId("self", "user")
-    ): MemberDetails.Self = mockk<MemberDetails.Self>().also {
-        every { it.selfUser } returns mockk<SelfUser>().also { user ->
-            every { user.id } returns id
-            every { user.name } returns name
-            every { user.previewPicture } returns null
-        }
-    }
-
-    private fun mockOtherUserDetails(
-        name: String,
-        id: UserId = UserId("other", "user"),
-        userType: UserType = UserType.INTERNAL
-    ): MemberDetails.Other = mockk<MemberDetails.Other>().also {
-        every { it.otherUser } returns mockk<OtherUser>().also { user ->
-            every { user.id } returns id
-            every { user.name } returns name
-            every { user.previewPicture } returns null
-        }
-        every { it.userType } returns userType
-    }
-
-    private fun mockUIMessage(userName: String = "mockUserName"): UIMessage {
-        return mockk<UIMessage>().also {
-            every { it.user } returns mockk<User>().also {
-                every { it.avatarAsset } returns null
-                every { it.availabilityStatus } returns UserStatus.AVAILABLE
-            }
-            every { it.messageSource } returns MessageSource.OtherUser
-            every { it.messageHeader } returns mockk<MessageHeader>().also {
-                every { it.messageId } returns "someId"
-                every { it.username } returns UIText.DynamicString(userName)
-                every { it.isLegalHold } returns false
-                every { it.time } returns ""
-                every { it.messageStatus } returns MessageStatus.Untouched
-            }
-            every { it.messageContent } returns null
-        }
-    }
-
-    private fun mockedMessage(senderId: UserId) = Message(
-        id = "messageID",
-        content = Text("Some Text Message"),
-        conversationId = ConversationId("convo-id", "convo.domain"),
-        date = "some-date",
-        senderUserId = senderId,
-        senderClientId = ClientId("client-id"),
-        status = Message.Status.SENT,
-        editStatus = Message.EditStatus.NotEdited
-    )
 }
